Add support for type annotations.
Type annotations are translated into applications of coercion
functions:

    e :: t   ==>   c_t e

    \(x :: t) -> e  ==> \x -> let x = (x :: t) in e

A coercion is built by translating the type into a graph and then
creating a function type:

    c_t :: forall (a = t) (b > t). a -> b

Note the rigid binding of `a`.  This has the effect that if `t` is
polymorphic, we *require* the polymorphism in the argument and *allow*
polymorphism in the result.

There are quite a number of TODOs left, but it passes basic tests.