Fix a bug in the graphic type -> syntact type translation.
The problem is that we calculate the context by starting from the root
and following the nodes bound under each node.  This means that if a
node is bound *above* the root we will not consider it.  That's fine
if the root is monomorphic in the given context, but it's wrong if the
node is bound at the gen-node directly above the root.

To be able to "see" these nodes we therefore add all the nodes that
are bound at the gen-node above the root to the root itself (taking
care that this does *not* include the root itself.)